feat(refactor): Restructure repo, add type safety, cross-chain compatible datasource interface (#255)
### Description

For
[ACT-788](https://linear.app/valora/issue/ACT-788/create-generic-interface-to-share-between-alchemy-and-blockscout).

This PR does a few inter-related things. The primary goal of this PR was
to introduce an interface that can be used by Alchemy as well as
Blockscout to create datasources for both. Along the way, I found it
necessary to add some rigor to the types we use throughout the repo, as
there were a lot of places where types were either not present, or
explicitly incorrect. Types have been mostly consolidated into a new
`types.ts` file. Additionally, in order to set us up a bit better for
multiple different sources of blockchain TX data, I've re-jiggered the
directory structure a bit to make things clearer.

The main idea for the interface is that transaction fetching takes 3
main steps:
* Fetching the raw transactions, and shaping them into _some_ internal
representation, `T`.
* Classifying the transactions (which requires access to internal fields
on `T`)
* Serializing the transactions (which requires access to internal fields
on `T`)

While the internal representation `T` is never revealed at the
boundaries of the interface, in order to separate the fetching process
into logical steps (i.e., functions, whose outputs are piped into the
next's input), we need to genericize the class. We specifically _do not_
attempt to have a unified `T` between Blockscout and whatever future
datasources we support (namely, Alchemy) since we don't want to attempt
to generalize the internal Blockscout transaction representation. We
already have a pattern for classification and serialization, but it
depends on the particular Blockscout transaction representation we
currently use. As such, the `TransactionClassifier` class has also been
made generic to match the new interface.

This PR does _not_ update the current `BlockscoutAPI` class to use this
new interface, though it should be relatively simple, since
`getTokenTransactionsV2` already closely mirrors the 3-step process
above. When adding a new datasource, the process will be:
* Figure out a desired TX representation `T`
* Add this TX representation to `TransactionInterface` in `types.ts`
* Add new classifiers/serialization logic under a new directory in
`/events`
* Subclass `BaseDataSource`, and implement the required methods
 
### Tested

Unit tested.

### Backwards compatibility

Yes.

### Related issues

- Fixes
[ACT-788](https://linear.app/valora/issue/ACT-788/create-generic-interface-to-share-between-alchemy-and-blockscout).