fix(types): sync TokenTransactionV2 types with wallet (#267)
### Description

fees and metadata are required for TokenTransfer type in wallet
    
fees is required for TokenExchange type in wallet

if we continue to leave them as optional here, we run the risk of type
errors at runtime when we let the field be undefined for some case where
the wallet actually uses them

### Tested

compiled typescript

### Backwards compatibility

yes. Some fields are newly marked required, but no field is newly marked
optional. I also verified that there isn't weird logic in the wallet
where undefined/null values are treated differently than empty lists for
this field.

### Related issues

na