feat(alchemy): TokenSent.getEvent (#271)
### Description

implemented TokenSent.getEvent

NOTE: punts on handling fees, which are paid for in Eth and probably
require a breaking change to the API of the service. I am punting on
this problem since it was not anticipated as part of this issue and
increases scope significantly. See issue
[here](https://linear.app/valora/issue/ACT-840/display-fees-correctly-when-paid-in-native-token)

### Tested

new unit tests

### Backwards compatibility

na

### Related issues

cc
https://linear.app/valora/issue/ACT-792/serialize-classified-alchemy-transactions