feat(alchemy): Add raw Alchemy transaction fetching (#269)
### Description

For
[ACT-789](https://linear.app/valora/issue/ACT-789/implement-raw-tx-fetching-from-alchemy).

Adds raw transaction fetching for the Alchemy datasource. Some
background to explain the fetching process I landed on:

Alchemy's `getAssetTransfers` endpoint uses a data model of _asset
transfers_, instead of whole blockchain transactions. A single asset
transfer consists of the transfer of a single type of asset from one
address to another; _blockchain transactions_ consist of any number of
such transfers. `getAssetTransfers` can be queried for transfers _from_
or _to_ an address. Since our goal is to aggregate transfers together
into _transactions_, the general idea is to call `getAssetTransfers`
twice, for transfers from and to the user's address, and aggregate
transfers corresponding to identical transaction hashes into a
transaction object. Transfers do not contain information about fees
(which, after mulling it over for a bit, actually makes decent sense to
me), so we augment the aggregated transfers with fee information by
fetching transaction receipts via Alchemy's `getTransactionReceipt` RPC
API.

Calling `getAssetTransfers` exactly once for each of transfers to and
from a user's address misses out on some edge cases; the results of
`getAssetTransfers` are paginated, and it is possible that all the
transfers for some transaction are _not_ contained in a single page
alone. For the sake of posterity, a "complete" approach where all edge
cases are avoided is as follows:
* Get a page of *to* transfers, starting at (and including) the block
specified by an afterCursor, if provided
* Get a page of *from* transfers, starting at (and including) the block
specified by an afterCursor, if provided
* Find the minimum block number of all transfers among *to* and *from*.
This gives us concrete bounds for all transfers we _must_ collect.
(`getAssetTransfers` results are ordered from most-to-least-recent)
* Fetch subsequent pages of *to* and *from* pages until we receive
transfers with block number less than the min block from above. This
guarantees that we've completely specified all transfers for blocks up
to and including the min block.
* Merge these transfers, sort by block, and prune based on
`MAX_RETURN_RESULTS`
* If pruning actually occurred, set afterCursor to the min block of the
pruned results, hasNextPage to `true`

The above is a "complete" approach, but comes at the cost of potentially
making many (often unnecessary) requests to Alchemy. Alchemy imposes
pretty restrictive limits on API throughput, and `getAssetTransfers` is
expensive in terms of throughput. The approach that this PR opts for is
as follows:
* Fetch a page of *to* transfers (1000 transfers), starting at (and
including) the block specified by an afterCursor, if provided
* Fetch a page of *from* transfers (1000 transfers), starting at (and
including) the block specified by an afterCursor, if provided
* Merge these transfers, sort by block, and prune based on
`MAX_RETURN_RESULTS`
* If pruning actually occurred, or if Alchemy has more pages available,
set afterCursor to the min block of the pruned results, hasNextPage to
`true`.

This approach has the advantage of only ever calling Alchemy twice (and
is generally simpler) but opens up some _exceedingly_ rare edge cases.
If >=1000 transfers for a user occur on a _single block_ (such that one
or more of the responses of `getAssetTransfers` contain transfers for
just a single block number), some aggregated transactions may be
incomplete/missing transfers.

On a related note, both approaches suffer from a similarly unlikely edge
case with regard to pagination. If more than `MAX_RETURN_RESULTS`
(currently 25) _transactions_ for a user exist for a _single block_,
`afterCursor` will be set to that block. Subsequent fetches using that
block as a `startCursor` will fetch transfers beginning with that block
(inclusive) and the same results will be returned to the user. This can
be avoided in most cases by encoding additional information into the
cursor objects.

To summarize, the edge cases we are vulnerable to with this approach
include:
* More than 1000 transfers, per-user, per-block
* More than `MAX_RETURN_RESULTS` (25) transactions, per-user, per-block

I really don't see either of these ever realistically happening (esp.
for actual, human-being Valora users), which is why I opted for this
approach in order to avoid rate limiting from Alchemy. As long as these
assumptions hold (that those two edge cases will not occur) subsequent
results returned from Alchemy by progressing the `afterCursor` will
always be unique, and contain full transactions only. If it's something
we're concerned about, it's not difficult to detect when these edge
cases are triggered and log a warning in order to track frequency.

Both Alfajores and Mainnet GCP have been updated with secrets containing
the Alchemy Ethereum API keys, and these have been added to the deploy
yamls.

### Tested

Unit and manual tested by running an instance of blockchain-api locally.

### Backwards compatibility

Yes.

### Related issues

- Fixes
[ACT-789](https://linear.app/valora/issue/ACT-789/implement-raw-tx-fetching-from-alchemy).